Change Public Suffix List helper packages
Switch from using the deprecated (and very out-of-date) publicsuffix
package to using the currently maintained publicsuffixlist package. The
publicsuffixlist package has a compatibility layer specifically for
code that is already using the publicsuffix package so changes are
minimal.

With pshtt switching to this same library it now handles all aspects of
loading the Public Suffix List internally. We only pass anything if we
want to use a specific file as the list source.